some changes to math functions and optimizations
faster pMultiply. also returning a 16 bit int or a 32 bit int depending on precision....
it is faster because it skips byte size checks and some sign checks and answer is a real number between +/- of a set value
because of some skipped checks it is faster than a normal long multiply. or so in theory...